+++ app/controllers/posts_controller.rb
@@ -23,6 +23,7 @@ class PostsController < ApplicationController
     @comments = @post.comments.includes(:user)
     @comment = Comment.new
     
+    
     respond_to do |format|
       format.html
       format.json { render json: { post: @post, comments: @comments } }
@@ -97,6 +98,7 @@ class PostsController < ApplicationController
   private
 
   def post_params
-    params.require(:post).permit(:title, :content, :image)
+    params.require(:post).permit(:title, :content, :image).merge(user_id: current_user&.id || User.guest.id)
   end
+  
 end
