+++ app/javascript/application.js
@@ -1,4 +1,5 @@
 import "@hotwired/turbo-rails"
 import "post"
 import "edit-delete-modal"
-import "comment-modal"
+import "comment"
+import "edit-delete-comment"
